My project "Ship Owners".
Realized: 1. Full Communication with server by JSON;
2. DTO Classes for Entities;
3. Using PostgreSQL DB and Hibernate;
4. Using Postman Desktop Application

The plans are to format the code and continue expanding the application.

+++ src/main/java/ru/shipownerproject/databases/vesselrepository/VesselRepository.java
@@ -7,7 +7,7 @@ import ru.shipownerproject.models.vessels.Vessel;
 import java.util.List;
 
 @Repository
-public interface VesselRepository extends JpaRepository<Vessel, Integer> {
+public interface VesselRepository extends JpaRepository<Vessel, Long> {
 
     List<Vessel> findByIMO(String IMO);
